backend/requirements.txt
fastapi
pydantic
pymongo
python-dotenv
typing-extensions
uvicorn[standard]
openai